# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Communications Microservice - (Twitter channel)",
        "description": "You need to write a REST API microservice. <br>\n    You need use Lumen + Swagger + Twitter SDK <br>\n    https://github.com/laravel/lumen.git <br>\n    https://github.com/DarkaOnLine/SwaggerLume <br>\n    https://github.com/atymic/twitter <br>\n    It is necessary to integrate Twitter SDK into our REST API microservice. <br>\n    We import one of the SDK to choose from those listed below. <br>\n    We implement all the logic as much as possible with the built-in functionality of Lumen",
        "contact": {
            "name": "Ojo Babajide Joshua",
            "url": "https://geenius.zyrocs.com",
            "email": "ojobabajide2018@gmail.com"
        },
        "license": {
            "name": "Your License",
            "url": "https://geenius.zyrocs.com"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/subscribe-to-chatbot": {
            "post": {
                "tags": [
                    "Subscription"
                ],
                "summary": "Subscribe users to a Twitter chatbot",
                "operationId": "subscribeToChatBot",
                "requestBody": {
                    "description": "User ID",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 123
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User subscribed to Twitter chatbot",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User subscribed to Twitter chatbot"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
    }
}